chore(pegasus): adapt install-on-chain to core bootstrap
 - consume pegasusBundle from the bootstrap promise space.
   - no need to exclude install-on-chain.js from lint, so move to src/
   - there's only one bundle, so never mind the Promise.all loop.
   - take care not to give Zoe.install a Promise<bundle>; that fails
     with: unrecognized moduleFormat ' undefined '
 - export CONTRACT_NAME = 'Pegasus' (capital P)
 - since we get nameAdmins from the promise space,
   use eventual send to call it.
 - never mind creatorFacet; Pegasus doesn't have one.
 - prune unused uiConfig
 - prune import of BootstrapPowers type
   ran into trouble with ambient types
